Update to latest generated specs
- Add remote identity id getter to CC context
- Reduce size of sp_id to 32 bits

+++ specs/c/tkm/types.h
@@ -28,7 +28,7 @@ typedef uint64_t eag_id_type;
 
 typedef uint64_t dhag_id_type;
 
-typedef uint64_t sp_id_type;
+typedef uint32_t sp_id_type;
 
 typedef uint64_t authp_id_type;
